Create brain class
brain includes basic move logic.

board gains width property for understanding boundaries & a function for
determining whether or not a coordinate is within bounds.

snake gains a method for to expose whether or not a coordinate is within
its body.

note: pylint is showing any typings in signatures as invalid syntax, so
I'm not currently using the typings until I can figure out how to fix
that.