WIP: Create brain test and update other tests
Brain still needs to be completed - but I'm unsure as to whether or not
we can continue with the current approach of not using other classes'
methods and stuff. Before we were avoiding it cause the tests are just
intended to test the one class.

Also, I added tests for new board and snake functions.