Accept risk of headons when there are no other move options
Still wanna avoid headon collisions in general, but there's a case which
I've seen several times now where the snake considered the
headon-collisions-risk-square an invalid move, so it goes towards its
tail as default because there are not valid_moves. but if it went
towards the headon in those particular cases, it could've survived.

from running a few games it seems rare enough that it only occurs like 4-5
 times in a 1500-turn game with 6 snakes. so not a huge issue, but also
worth avoiding in my opinion.