Fix start coord for path from food to tail
in the path-getter, we verify the path to the coord by ensuring there's
a path from that coord to our tail, otherwise the coord isn't an
acceptible destination because we'd get stuck.

the "start" coordinate for the path back from the destination to the
tail wasn't updating after advancing the snake along the path, and
so the function wasn't actually serving its purpose.

it wouldn't update because the reference to the new board with the
updated snake state (advanced along the to-food path) was not the source
of the start coordinate.